Alterações solicitadas após teste

+++ app/Http/Controllers/FornecedorController.php
@@ -26,7 +26,7 @@ class FornecedorController extends Controller
     {
         if(Auth::check()){
             if(Auth::user()->type == 2 || Auth::user()->type == 1){
-                $fornecedores = Fornecedor::withTrashed()->paginate(10);
+                $fornecedores = Fornecedor::orderBy('nome')->withTrashed()->paginate(10);
                 return view('fornecedor.index')->with('fornecedores', $fornecedores);
             }else{
                 return redirect()->back();
@@ -63,11 +63,35 @@ class FornecedorController extends Controller
      */
     public function store(Request $request)
     {
-        Fornecedor::create($request->all());
-        $request->session()->flash('success', 'Fornecedor inserido com Sucesso!');
-                return redirect()->route('fornecedor.index');
-    }
+         if(Auth::check()){
+            if(Auth::user()->type == 2 || Auth::user()->type == 1){
 
+                 $mensagens = ['nome.required' => 'O campo :attribute é obrigatório',
+                'email.required' => 'O campo :attribute é obrigatório',
+                'telefone.required' => 'O campo :attribute é obrigatório',
+                'descricao.required' => 'O campo :attribute é obrigatório',
+                'cnpj.required' => 'Esse :attribute já foi cadastrado',
+                'cnpj.unique' => 'Esse :attribute já foi cadastrado',
+                ];
+                $request->validate([
+                    'descricao'=>'required',
+                    'nome'=>'required',
+                    'email'=>'required|email',
+                    'telefone'=>'required',
+                    'cnpj'=>'required|unique:fornecedores',
+                ], $mensagens);
+
+                Fornecedor::create($request->all());
+                $request->session()->flash('success', 'Fornecedor inserido com Sucesso!');
+                 return redirect()->route('fornecedor.index');
+
+            }else{
+                return redirect()->back();
+            }  
+        }else{
+            return redirect()->back();
+            } 
+    }
     /**
      * Display the specified resource.
      *
@@ -116,13 +140,38 @@ class FornecedorController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Fornecedor $fornecedor)
-    {
-        $fornecedor->fill($request->all());
-        $fornecedor->save();
-        session()->flash('success', 'Fornecedor atualizado com sucesso!');
-        return redirect()->route('fornecedor.show', $fornecedor->id);
-    }
+    { 
 
+        if(Auth::check()){
+            if(Auth::user()->type == 2 || Auth::user()->type == 1){
+
+            $mensagens = ['nome.required' => 'O campo :attribute é obrigatório',
+            'email.required' => 'O campo :attribute é obrigatório',
+            'telefone.required' => 'O campo :attribute é obrigatório',
+            'descricao.required' => 'O campo :attribute é obrigatório',
+            'cnpj.required' => 'Esse :attribute já foi cadastrado',
+            'cnpj.unique' => 'Esse :attribute já foi cadastrado',
+            ];
+            $request->validate([
+                'descricao'=>'required',
+                'nome'=>'required',
+                'email'=>'required|email',
+                'telefone'=>'required',
+                'cnpj'=>'required|unique:fornecedores',
+            ], $mensagens);
+
+            $fornecedor->fill($request->all());
+            $fornecedor->save();
+            session()->flash('success', 'Fornecedor atualizado com sucesso!');
+            return redirect()->route('fornecedor.show', $fornecedor->id);
+            }else{
+                return redirect()->back();
+                }
+       
+        } else{
+            return redirect()->back();
+            } 
+    }
     /**
      * Remove the specified resource from storage.
      *
@@ -131,17 +180,33 @@ class FornecedorController extends Controller
      */
     public function destroy(Fornecedor $fornecedor)
     {
-        $fornecedor->delete();
-        session()->flash('success', 'Fornecedor deletado com sucesso!');
-        return redirect()->route('fornecedor.index');
+        if(Auth::check()){
+            if(Auth::user()->type == 2 || Auth::user()->type == 1){
+                $fornecedor->delete();
+                session()->flash('success', 'Fornecedor deletado com sucesso!');
+                return redirect()->route('fornecedor.index');
+            }else{
+                return redirect()->back();
+                }
+        }else{
+            return redirect()->back();
+            } 
     }
 
     public function restore($id)
     {  
+        if(Auth::check()){
+            if(Auth::user()->type == 2 || Auth::user()->type == 1){
 
-        Fornecedor::onlyTrashed()->where('id', $id)->restore();
-        session()->flash('success', 'Fornecedor ativado com sucesso!');
+                Fornecedor::onlyTrashed()->where('id', $id)->restore();
+                session()->flash('success', 'Fornecedor ativado com sucesso!');
+                return redirect()->route('fornecedor.index');
 
-        return redirect()->route('fornecedor.index');
+            }else{
+                return redirect()->back();
+                }
+        }else{
+            return redirect()->back();
+        } 
     }
 }
